requirements.txt
langchain==0.1.16
langchain-core==0.1.27
langchain-groq==0.0.6
langgraph==0.0.27
python-dotenv==1.0.0
arxiv==2.1.0
wikipedia==1.4.0
tavily-python==0.3.1 